GHC 8 compat: Semigroup is a superclass of Monoid
This fails compiling with recent versions of GHC due to a missing
Semigroup instance for Stream. But at the same time, we don't want to
pull in the `semigroups` package for older GHC - so just make the
instance conditional for versions of base which contain a definition of
Data.Semigroup.

base-4.11 is the first version that exports Semigroup from Prelude, and
also the first version that makes Semigroup a super-class of Monoid.